+++ oshdb/src/test/java/org/heigit/ohsome/oshdb/grid/GridOSHRelationsTest.java
@@ -27,9 +27,9 @@ public class GridOSHRelationsTest {
 
   @Test
   public void test() throws IOException {
-    var node100 = buildOSHNode(node(100L, 1, 1L, 0L, 123, tags(1, 2), 494094984L, 86809727L));
-    var node102 = buildOSHNode(node(102L, 1, 1L, 0L, 123, tags(2, 1), 494094984L, 86809727L));
-    var node104 = buildOSHNode(node(104L, 1, 1L, 0L, 123, tags(2, 4), 494094984L, 86809727L));
+    var node100 = buildOSHNode(node(100L, 1, 1L, 0L, 123, tags(1, 2), 494094984, 86809727));
+    var node102 = buildOSHNode(node(102L, 1, 1L, 0L, 123, tags(2, 1), 494094984, 86809727));
+    var node104 = buildOSHNode(node(104L, 1, 1L, 0L, 123, tags(2, 4), 494094984, 86809727));
 
     var way200 = buildOSHWay(asList(node100, node104),
             way(200, 1, 3333L, 4444L, 23, tags(1, 2), mn(100, 0), mn(104, 0)));
@@ -65,7 +65,7 @@ public class GridOSHRelationsTest {
   }
 
   private static OSMNode node(long id, int version, long timestamp, long changeset,
-      int userId, int[] tags, long longitude, long latitude) {
+      int userId, int[] tags, int longitude, int latitude) {
     return new OSMNode(id, version, timestamp, changeset, userId, tags, longitude, latitude);
   }
